+++ MyLittleUI.cs
@@ -23,7 +23,7 @@ namespace MyLittleUI
     {
         const string pluginID = "shudnal.MyLittleUI";
         const string pluginName = "My Little UI";
-        const string pluginVersion = "1.1.11";
+        const string pluginVersion = "1.1.12";
 
         private readonly Harmony harmony = new Harmony(pluginID);